+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v0.3.3
+- Moved Suspicious Tentacle compatibility config out of BodyBlend settings.
+- Made Suspicious Tentacle config easier to use.
+- Suspicious Tentacle now animates BodyBlend more smoothly.
+- Added API for other mods to access all the loaded parts names. Mostly used for setting up configuration file.
+
+v0.3.2
+- Added config for setting minimum size.
+
 v0.3.1
 - Fix issue with dynamic bone values config.